web-runtime: Update from chromium68 to chromium79
This patch contains all the squashed changes needed
to bump WAM and Chromium to M79, including all the
ported changes from both upstream Chromium, webOSE and
and the former chromium 68 version.

It includes the needed changes on the recipes, and bumps
the versions to the new commits/branches of the
Chromium and Wam repositories.

The split of the ongoing changes merged is available here:
https://github.com/Igalia/meta-agl-html5-dev/commits/chromium79

Bug-AGL: SPEC-3576

Change-Id: Ia0de6f11001189789d4ffbf01ed81929c57bc7c5
Signed-off-by: Lorenzo Tilve <ltilve@igalia.com>