packagegroup-agl-image-ivi: restore agl-login-manager dependency
Restore the agl-login-manager RDEPENDS to packagegroup-agl-image-ivi
so that it will get pulled into agl-image-ivi(-crosssdk) images.
This was removed during the layer reorganization done for SPEC-3723,
but is required so that afm-user-session and its dependencies will be
started for boot.  This will restore CI functionality while further
post layer reorganization cleanup to rationalize where the dependency
should go is considered.

Bug-AGL: SPEC-3753

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I2ef8a5bf05e5c29267643ca93748ae7dad42c8fa