Add dev-mapping-ci recipe
Add a dev-mapping-ci recipe that installs a CI specific version of
/etc/dev-mapping.conf for the agl-service-can-low-level binding.
Additionally, the dev-mapping-demo recipe files have been moved to
recipes-config/dev-mapping so all the dev-mapping* recipes are in
the same place, and the now required "j1939" bus mapping has been
added to the various dev-mapping files.

To enable use of the new recipe, a definition of the variable
PREFERRED_RPROVIDER_virtual/low-can-dev-mapping has been added to
the agl-demo feature definition that enables it based on the AGLCI
DISTRO_FEATURE.  This is intended to replace the definition in
poky-agl.conf, which will be removed in subsequent cleanup.

Bug-AGL: SPEC-3756

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ieb80a7f4b76ddd7851acca521dea1a8dd104cf86