web-runtime: Bump chromium79 and wam to adress first-frame dection and logging
Bump the versions of Chromium79 to include a workaround for the
webapps launching delay that was introduced in order to address
problems happening with first-frame detection.

These updates also fix full WAM verbosity to journalctl when defined,
and building Chromium with externalsrc, and removes lld and gold.

Also iclude explicit dependency chain when building html5 dempo
platform to packagegroup-agl-profile-graphical-html5.

Bug-AGL: SPEC-3576
Bug-AGL: SPEC-3767

Signed-off-by: Lorenzo Tilve <ltilve@igalia.com>
Change-Id: I34b0001c58522cfc7940d0384809e14e3d999ea8