wam: Fix startup issue on qemux86-64
There was an issue on WAM startup related with the initialization
of the pmlog that was causing a crash on qemu. Update the WAM version
to avoid that crash.

It includes a revert on a change added to the logging system that
triggered memory errors on qemu, and which with the
fix for user-data-dir creation, allow WAM to startup correctly.

https://github.com/Igalia/webosose-wam/commit/8a1ec2f3e474199a689ae71a12ba70a230770c0d

BUG-AGL: SPEC-3780
BUG-AGL: SPEC-3781

Signed-off-by: Lorenzo Tilve <ltilve@igalia.com>
Change-Id: Id528f8e3270e4f5417fbfda4cdb6a10b001cd6bb