Update cluster demo weston configuration
Tweak the cluster demo to move the weston-ini-conf-landscape
dependency to the image recipe, where the use of the regular or
inverted configuration is now based on whether the "agl-demo-preload"
feature is set.  This means the cluster demo images will now by
default have no rotation, but still allows building images for the
AGL demo setup with its inverted cluster display.

Bug-AGL: SPEC-3791

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I4f82cc84d27687f228ed3b9587103edef9c150c0