cluster-demo-simulator: update LICENSE/LIC_CHKSUM_FILE
Upstream has renamed the common license files, but investigation
revealed that the LICENSE on this purely configuration recipe
should likely be MIT to be consistent with the other similar
recipes in the tree, so update both LICENSE and LIC_CHKSUM_FILE
to fix both issues.  As well, now that the demo repositories
are merged, this recipe can just RDEPEND on simple-can-simulator,
rather than duplicating it.

Bug-AGL: SPEC-3819

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I2302e904258a67d3c2a1756f94d247e3779106f0