Add kernel-modules package to demo images
Add the kernel-modules meta package to demo images in the agl-demo.inc
aglsetup.sh template.  This change is to balance out a removal done
in the base template in meta-agl and retain the build of all kernel
modules into the demo platform images for developer convenience.

Note that the inclusion is conditional on the virtual/kernel provider
not being linux-dummy, as it is desirable to leave even the dummy
kernel-modules package out of container images to avoid confusion.

Bug-AGL: SPEC-3983

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I3c4e3876c11d7fb3f1796841aad8a9a463d8e513