wam_git: Add protobuf dependency for WAM
We're not currently using protobuf, but it would be of tremendous help
if we are to be ready for the next integration phase which started to
make use of it. Makes development much easier to do and it won't make
such a big impact to have it early.

Bug-AGL: SPEC-3967

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Change-Id: Ie35204a11a1f2e28c4fa3ce6566d17cc8c0421dd