chromium84: Add a delay to agl ready event
Delay ready() to ensure that all the window and surfaces setup
is done before notifying the compositor

There's a work in progress for enabling wam to display multiple
surfaces for a single widget, which would eliminate the chain
loading that currently causes this issue.

This change includes a temporary patch that should be removed
once wam changes are integrated.

Bug-AGL: SPEC-3799

Signed-off-by: Roger Zanoni <rzanoni@igalia.com>
Change-Id: I9945af60077a79d56052c3fdb358e9fed44874e6