Convert to new override syntax
This is the result of running a slightly customized version of the
convert-overrides.py script from poky with additional overrides
added.

The intent of these changes is to minimize the effort to keep the
"next" branch that builds against poky master up to date and tested
in preparation for the switch to the next Yocto LTS release in
early 2022.

Bug-AGL: SPEC-4052

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I6eeadd2c985c008b04f8f9815bf0f6fb064bedac