agl-users: remove display user
The "display" user has effectively been replaced with a "weston"
user in the new weston-init support for running as non-root from
upstream Yocto, so remove creation of it from the agl-users recipe.
If the agl-users recipe will be retained in the new application
scheme, this cleanup is required.

Bug-AGL: SPEC-4121

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Iafd7a08dfd7ba3514ef4264c68d697cfd9790047