Add updated demo applications back
Add updated recipes for the dashboard, hvac, ondemandnavi, and
settings Qt demo applications, as well as for the libqtappfw library
they use and its new dependency bluez-glib.  The applications are
added back to the packagegroup-agl-demo-platform as they were
before.  Also remove a couple of leftover things in the packagegroup
recipe to clean up.

Bug-AGL: SPEC-4182

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: I0ea36c809d9126284c597cd79eded8ab031205da