Add backports of mpd and related recipes
Backport the recipe for the latest (0.23.5) version of mpd
(Music Player Daemon) from meta-openembedded's master branch.
This provides a version with native PipeWire support, and should
help make the upcoming upgrade to Yocto Project 3.5 (aka Kirkstone)
easier.  Additionally, the recipes for the related libmpdclient
library and mpc utility are also backported to ensure compatibility.

The provided bbappend for the mpd recipe tweaks the backport to
build with the older version of meson in dunfell, and applies
configuration to approximate the lightmediascanner set up with
regards to media location and using /var/volatile for metadata
storage for robustness.

Bug-AGL: SPEC-4255

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Ida58289a16079f48c3467d55aef4a608671ffc37