libqtappfw: enable building mpd based mediaplayer support
Update libqtappfw recipe to pull in libmpdclient dependency
for building the reworked mediaplayer support, and also add
a RRECOMMENDS definition that pulls in bluez5, connman, and
mpd to make the relationship more explicit.  The intent is
to split the libqtappfw-X libraries into separate packages
with some follow-on rework and have these dependencies be
more fine-grained at that point.

Bug-AGL: SPEC-4182

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: If7fcd57dab89a22dfee81017d0411854b1f16bda