wam: improve shutdown procedure
It was observed that restarting agl-session would crash and/or hang
WAM. So this change binds WAM to agl-compositor in systemd service
file. This means stopping compositor will stop WAM.

It also fixes another problem on restart, as the WAM launcher
processes from previous execution would still be alive. So it forces
killing all the WAM processes for the user on shutdown.

Bug-AGL: SPEC-4293
Change-Id: I5131cdd56c292c66d0f2d5ea88cf772ef9b4bd56
Signed-off-by: Jose Dapena Paz <jdapena@igalia.com>