html5-hvac, html5-homescreen, html5-launcher: use explicit GIT hashes
To allow coordination of landing of features in wam and chromium with
the web application changes, add explicit hashes.

This is required to later land the changes for appinfo.json, as web
applications will need to change at the same time WAM stops parsing
config.xml and starts using appinfo.json.

Bug-AGL: SPEC-4251
Change-Id: Ib1f3af7f9919d467bd7ade2dd6166c4242a8316c
Signed-off-by: Jose Dapena Paz <jdapena@igalia.com>