Remove use of AGL_APP_REVISION
We did use the variable AGL_APP_REVISION to change between the HEAD of the branch
via  to the fixed tag at release time (regs/tags/xyz) .

While this worked well previously, it turns out that bitbake will query git every time
it runs for either  or tags likewise. For tags it cannot trust whatever is known locally
is actually the latest tags as there can be force pushed tags.

To fix this we use a well-defined SRCREV for apps/* in the same way
as we do it for src/* .

Bug-AGL: SPEC-4455
Signed-off-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Change-Id: I97fc125d40b36b7ac5e069018a8973384f61d0b7