Add agl-ivi-demo-platform-flutter image
Add an image for testing Flutter applications.  At present it uses
the Qt demo homescreen and launcher until they can be replaced with
a Flutter homescreen with equivalent functionality.  To allow
testing agl-compositor behavior, the packagegroup for the image
includes the demo Qt HVAC and dashboard apps in addition to the
Flutter "gallery" demo application packaged for launching with
applaunchd.  The "agl-flutter" build feature has been added as a
dependency of "agl-demo" in order for the above to be buildable by
default.

Bug-AGL: SPEC-4485

Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Change-Id: Idf2daa189a15e9cc4bfe2030ac6aa3f5b0331c0c