native-shell-client: Add systemd unit
Add a bbappend for native-shell-client to add a systemd unit that
runs in the agl-driver user session.  This is required to actually
get it started if it is built into an image.

Bug-AGL: SPEC-4618

Change-Id: If6c1c015544ee6f126fb30ca67822a38a83df64c
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
(cherry picked from commit 2848a3d0591c7d0c9e97f208f00899a0f254ad0a)