linux-%: Tweak kernel configuration for KVM demo guests
Rework the kernel configuration fragment application to prune out
a bunch of the configuration for physical hardware for the
virtio-aarch64 kernel.  As well, add a virtio guest kernel specific
fragment to enable a PCI CAN device driver that works with QEMU's
emulation.  That will provide a fallback until something can be
worked out with respect to virtio-can.

Bug-AGL: SPEC-4618

Change-Id: Ibb674d09c115ab5bbe454b665350a658c759e5f5
Signed-off-by: Scott Murray <scott.murray@konsulko.com>