output-udev-conf: add recipe
Add recipe to install udev rules to force all input devices to be
associated with the HDMI-A-1 output.  The intended use is with the
KVM host image to ensure input devices are matched to the IVI
guest VM.  Also fixed the RDEPENDS for the other udev rule recipes
to depend on udev specifically instead of systemd.

Bug-AGL: SPEC-4618

Change-Id: Ic2efe01f7aed85f89dbda3cb958332ed5db799a1
Signed-off-by: Scott Murray <scott.murray@konsulko.com>