agl-qemu-runner: Add recipes
Add recipes for agl-qemu-runner.sh QEMU wrapper script and template
systemd unit, and guest per-image configuration files and systemd
unit links for the various demo platform images.  The configurations
for the demo images can be mixed and matched somewhat, but only one
each of IVI and cluster should be used, as there are some static CPU
assignments in the configurations.  As well, this has only been
tested on the AGL reference hardware at present, some rework will
likely be required to support other platforms.

There is room to add a bbclass to simplify the guest configuration
recipes, but that has not been done for now as this approach may be
temporary; a switch to libvirtd with appropriate configurations is
likely.

Bug-AGL: SPEC-4618

Change-Id: Ieefbd4081cca506f915957cc313411a5ed1ece5c
Signed-off-by: Scott Murray <scott.murray@konsulko.com>