agl-kvm-demo-platform: Add recipe
Add agl-kvm-demo-platform image recipe.  If "agl-kvm" has been given
to aglsetup.sh, it can be built to generate a minimal host image that
runs IVI and cluster images with QEMU+KVM.  The guest images are
built with the "agl-guest-kvm" that the "agl-kvm" feature enables,
and are built against the virtio-aarch64 target.  The guest images
to build can be overriden with the variables GUEST_VM1_IMAGE and
GUEST_VM2_IMAGE (from e.g. site.conf).

At present this image has only been tested on the AGL reference
platform, and development is likely required to get it working on
other platforms.  COMPATIBLE_MACHINE has not been set in the recipe
to ease experimenting, that may change.

Known issues:
- Currently graphics output from the guests is corrupted when they
  first boot, and a manual reboot is required of each guest.
- Peripheral support is still a WIP.  Audio has been configured,
  but needs further debugging.  CAN, USB, etc., are to follow.

Bug-AGL: SPEC-4618

Change-Id: Id7113576453621702547ee6e0b8734a6e9117bb6
Signed-off-by: Scott Murray <scott.murray@konsulko.com>